citation/requirements.txt
pika
psycopg2
python-dotenv
google-cloud-storage
refextract